Refactor sidebar logic for logged-in users

+++ home.py
@@ -53,28 +53,28 @@ with col2:
                     st.session_state.selected_grade = int(grade)
                     st.write("Grade " + str(st.session_state.selected_grade))
                     st.rerun()
-
-s1, s2 = st.sidebar.columns(2, vertical_alignment="center")
-with s1:
-    if st.session_state.streak: # add reset streak and incomplete streak
-        with open(f"{st.user.email}streak.txt", "r+") as f:
-            f.seek(7)
-            z = int(f.read()) + 1
-            f.truncate()
-            f.write(f"{z}")
-        st.markdown(":color[:material/mode_heat:" + z + "]{foreground='#fa3002'}")
-        st.session_state.streak = False
-with s2:
-    @st.dialog("Test Score", dismissible=True)
-        def show_popup_message(message_text):
-            st.write(message_text)
-        with open(f"{st.user.email}score.txt", "r+") as f:
-            f.seek(6)
-            w = int(f.read()) + st.session_state.score
-            f.truncate()
-            f.write(f"{w}")
-        if st.button(f"{w}", type="tertiary", use_container_width=True):
-            show_popup_message("The test score is used to define your overall improvements. The more positive feedback you receive, the higher your score goes, and the more negative feedback you receive, the lower your score goes. Your streak also affects your test score. If your streak is higher, your score will increase faster, while reducing the amount you decrease when given negative feedback. Complete a test every day (including diagnostics) to increase your streak. If you forget to test for just one day, your streak restarts at 0.")
+if st.user.is_logged_in:
+    s1, s2 = st.sidebar.columns(2, vertical_alignment="center")
+    with s1:
+        if st.session_state.streak: # add reset streak and incomplete streak
+            with open(f"{st.user.email}streak.txt", "r+") as f:
+                f.seek(7)
+                z = int(f.read()) + 1
+                f.truncate()
+                f.write(f"{z}")
+            st.markdown(":color[:material/mode_heat:" + z + "]{foreground='#fa3002'}")
+            st.session_state.streak = False
+    with s2:
+        @st.dialog("Test Score", dismissible=True)
+            def show_popup_message(message_text):
+                st.write(message_text)
+            with open(f"{st.user.email}score.txt", "r+") as f:
+                f.seek(6)
+                w = int(f.read()) + st.session_state.score
+                f.truncate()
+                f.write(f"{w}")
+            if st.button(f"{w}", type="tertiary", use_container_width=True):
+                show_popup_message("The test score is used to define your overall improvements. The more positive feedback you receive, the higher your score goes, and the more negative feedback you receive, the lower your score goes. Your streak also affects your test score. If your streak is higher, your score will increase faster, while reducing the amount you decrease when given negative feedback. Complete a test every day (including diagnostics) to increase your streak. If you forget to test for just one day, your streak restarts at 0.")
 with st.sidebar:
     if st.session_state.get("guest_mode") and st.session_state.get("guest_login_time"):
         start_time = st.session_state.guest_login_time
